[BE] feat(game): 게임 세션 스키마와 월드·대출 연동 구조를 반영한다 (S14P21C103-196)
- GameSession aggregate와 상태 enum, GameHousing ownership 이동을 반영한다
- GameSessionRef와 대출 provider 추상화를 제거하고 session 모델 기준으로 연동을 전환한다
- world, realestate, loan 경로와 schema.sql, dummy 데이터를 새 구조에 맞춘다
- repository 테스트를 SpringBootTest 기준으로 정비한다

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/controller/game/loan/LoanController.java
@@ -43,7 +43,7 @@ public class LoanController {
      */
     @GetMapping("/products")
     public ApiResponse<List<LoanProductResponse>> getProducts(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @RequestParam(defaultValue = "ALL") final String category,
             @RequestParam(defaultValue = "0") final int page,
             @RequestParam(defaultValue = "20") final int size
@@ -58,7 +58,7 @@ public class LoanController {
      */
     @GetMapping("/products/{productId}")
     public ApiResponse<LoanProductDetailResponse> getProductDetail(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @PathVariable final String productId
     ) {
         final LoanProductDetailResponse detail = loanProductService.getProductDetail(productId)
@@ -72,7 +72,7 @@ public class LoanController {
      */
     @PostMapping("/calculate")
     public ApiResponse<LoanCalculateResponse> calculate(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @RequestBody final LoanCalculateRequest request
     ) {
         final LoanCalculateResponse response = loanService.calculate(
@@ -87,7 +87,7 @@ public class LoanController {
      */
     @PostMapping("/apply")
     public ApiResponse<LoanApplyResponse> apply(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @RequestBody final LoanApplyRequest request
     ) {
         final LoanApplyResponse response = loanService.apply(
@@ -101,7 +101,7 @@ public class LoanController {
      */
     @PostMapping("/confirm")
     public ApiResponse<LoanConfirmResponse> confirm(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @RequestBody final LoanConfirmRequest request
     ) {
         final LoanConfirmResponse response = loanService.confirm(
@@ -116,7 +116,7 @@ public class LoanController {
      */
     @PostMapping("/repay")
     public ApiResponse<LoanRepayResponse> repay(
-            @PathVariable final Integer sessionId,
+            @PathVariable final Long sessionId,
             @RequestBody final LoanRepayRequest request
     ) {
         final LoanRepayResponse response = loanService.repay(
