+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/gamesession/stock/StockHoldingRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface StockHoldingRepository
         extends JpaRepository<StockHolding, StockHoldingId> {
 
-    List<StockHolding> findAllByGameSessionId(Integer gameSessionId);
+    List<StockHolding> findAllByGameSessionId(Long gameSessionId);
 }
