+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/GameContractReviewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface GameContractReviewRepository extends JpaRepository<GameContractReview, Long> {
 
     Optional<GameContractReview> findTopByGameSessionIdAndPropertyIdOrderByReviewedAtDesc(
-        Integer gameSessionId,
-        Integer propertyId
+        Long gameSessionId,
+        Long propertyId
     );
 }
